+++ src/main/java/com/sparta/tdd/domain/address/dto/naver/Meta.java
@@ -1,4 +1,4 @@
-package com.sparta.tdd.domain.address.dto;
+package com.sparta.tdd.domain.address.dto.naver;
 
 public record Meta(
         int totalCount,
