+++ src/main/java/com/sparta/tdd/domain/address/dto/naver/NaverAddressElement.java
@@ -1,4 +1,4 @@
-package com.sparta.tdd.domain.address.dto;
+package com.sparta.tdd.domain.address.dto.naver;
 
 import java.util.List;
